pyproject.toml
[project]
name = "fast-api-auth"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.5",
    "argon2-cffi>=25.1.0",
    "asyncpg>=0.30.0",
    "bcrypt>=5.0.0",
    "fastapi>=0.118.3",
    "passlib>=1.7.4",
    "psycopg2>=2.9.11",
    "pydantic-settings>=2.11.0",
    "pydantic[email]>=2.12.0",
    "python-jose[cryptography]>=3.5.0",
    "resend>=2.16.0",
    "uuid-utils>=0.11.1",
    "uvicorn[standard]>=0.37.0",
]
